split epoch API out of runtime adapter
