+++ chain/epoch-manager/Cargo.toml
@@ -22,6 +22,7 @@ chrono = { version = "0.4.4", optional = true}
 
 near-crypto = { path = "../../core/crypto" }
 near-primitives = { path = "../../core/primitives" }
+near-chain-primitives = { path = "../chain-primitives" }
 near-store = { path = "../../core/store" }
 near-chain-configs = { path = "../../core/chain-configs" }
 near-cache = { path = "../../utils/near-cache" }
