+++ tools/state-viewer/src/commands.rs
@@ -7,8 +7,9 @@ use ansi_term::Color::Red;
 use near_chain::chain::collect_receipts_from_response;
 use near_chain::migrations::check_if_block_is_first_with_chunk_of_version;
 use near_chain::types::{ApplyTransactionResult, BlockHeaderInfo};
+use near_chain::types::{EpochManagerAdapter, RuntimeAdapter};
 use near_chain::Error;
-use near_chain::{ChainStore, ChainStoreAccess, ChainStoreUpdate, RuntimeAdapter};
+use near_chain::{ChainStore, ChainStoreAccess, ChainStoreUpdate};
 use near_chain_configs::GenesisChangeConfig;
 use near_epoch_manager::EpochManager;
 use near_network::iter_peers_from_store;
